app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
chromadb-client==1.3.3
PyYAML==6.0.2